requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# there might be something here is useless but i dont remember which one, sorry
_tflow_select=2.1.0=gpu
absl-py=0.12.0=py37haa95532_0
astor=0.8.1=py37haa95532_0
blas=1.0=mkl
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2021.7.5=haa95532_1
certifi=2021.5.30=py37haa95532_0
cffi=1.14.5=py37hcd4344a_0
chardet=3.0.4=py37haa95532_1003
coverage=5.5=py37h2bbff1b_2
cryptography=3.4.7=py37h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cython=0.29.23=py37hd77b12b_0
freeglut=3.0.0=h6538335_5
freetype=2.10.4=hd328e21_0
gast=0.4.0=py_0
grpcio=1.36.1=py37hc60d5dd_1
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h6c2663c_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37haa95532_0
intel-openmp=2021.2.0=haa95532_616
jasper=2.0.14=hdc05fd1_1
joblib=0.17.0=py_0
jpeg=9d=h8ffe710_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libclang=11.1.0=default_h5c34c98_1
liblapack=3.8.0=20_mkl
liblapacke=3.8.0=20_mkl
libopencv=4.5.2=py37hebf5d07_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.15.8=h7755175_0
libtiff=4.2.0=hd0e1b90_0
libwebp-base=1.2.0=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.4=py37haa95532_0
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.3.0=py37h46781fe_0
mkl_random=1.1.1=py37h47e9c7a_0
mock=4.0.3=pyhd3eb1b0_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
olefile=0.46=py37_0
opencv=4.5.2=py37h03978a9_0
openssl=1.1.1k=h2bbff1b_0
pillow=8.0.0=py37hca74424_0
pip=21.1.1=py37haa95532_0
protobuf=3.15.8=py37hf2a7229_0
py-opencv=4.5.2=py37h7813e69_0
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_1
python=3.7.10=h6244533_0
python_abi=3.7=1_cp37m
qt=5.12.9=h5909a2a_4
requests=2.24.0=py_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.1.0=py37h4f6bf74_1
setuptools=52.0.0=py37haa95532_0
six=1.15.0=py37haa95532_0
sqlite=3.35.4=h2bbff1b_0
tensorboard=1.13.1=py37h33f27b4_0
tensorflow=1.13.1=h21ff451_7
tensorflow-base=1.13.1=py37_7
tensorflow-estimator=1.13.0=py_0
tensorflow-gpu=1.13.1=h0d30ee6_0
termcolor=1.1.0=py37haa95532_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.25.11=py_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
